rename useLink to useRouter + add router.back() and router.replace()
